requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis==5.0.1
google-generativeai==0.3.2
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
